pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-kube-architect-gym"
version = "0.1.0"
description = "OpenEnv environment for constrained Kubernetes architecture generation."
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git@v0.2.1",
    "openai>=1.0.0",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
train = [
    "accelerate>=1.1.0",
    "datasets>=3.0.0",
    "torch>=2.8.0,<2.10.0",
    "transformers>=4.55.0",
    "trl[vllm]==0.29.0",
    "vllm==0.11.2",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "kube_sre_gym.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["kube_sre_gym", "kube_sre_gym.server"]
package-dir = { "kube_sre_gym" = ".", "kube_sre_gym.server" = "server" }
